refactor: move data element to es_dao.go
