--- golang_api_server/pkg/services/news_doc_service/news_doc.go
@@ -1,15 +1,0 @@
-package news_doc_service
-
-import "time"
-
-type NewsDoc struct {
-	UUID        string     `json:"uuid"`
-	Link        string     `json:"link,omitempty"`
-	Title       string     `json:"title,omitempty"`
-	Description string     `json:"description,omitempty"`
-	Authors     []string   `json:"authors,omitempty"`
-	Category    string     `json:"category,omitempty"`
-	CreatedAt   *time.Time `json:"created_at,omitempty"`
-	UpdatedAt   *time.Time `json:"updated_at,omitempty"`
-	DeletedAt   *time.Time `json:"deleted_at,omitempty"`
-}
